+++ inc/abstractquestionparameter.class.php
@@ -85,6 +85,12 @@ PluginFormcreatorTranslatableInterface
       return $input;
    }
 
+   public function prepareInputForUpdate($input) {
+      unset($input['uuid']);
+
+      return $input;
+   }
+
    public function rawSearchOptions() {
       $tab = [];
       $tab[] = [
