+++ inc/questionfilter.class.php
@@ -104,7 +104,7 @@ extends PluginFormcreatorAbstractQuestionParameter
    }
 
    public function post_getEmpty() {
-      $this->fields['filter'] = '[]';
+      $this->fields['filter'] = [];
    }
 
    public function post_getFromDB() {
