+++ setup.php
@@ -40,7 +40,7 @@ define('PLUGIN_FORMCREATOR_SCHEMA_VERSION', '2.14');
 define('PLUGIN_FORMCREATOR_IS_OFFICIAL_RELEASE', false);
 
 // Minimal GLPI version, inclusive
-define ('PLUGIN_FORMCREATOR_GLPI_MIN_VERSION', '10.0');
+define ('PLUGIN_FORMCREATOR_GLPI_MIN_VERSION', '10.0.3');
 // Maximum GLPI version, exclusive (ignored if PLUGIN_FORMCREATOR_IS_OFFICIAL_RELEASE == false)
 define ('PLUGIN_FORMCREATOR_GLPI_MAX_VERSION', '10.1');
